multimodal-generation/sample-apps/01-aws-summit-seoul-2025-comicai-studio/tooncraft/requirements.txt
boto3>=1.26.0
requests>=2.26.0
requests-aws4auth>=1.3.0
pillow==11.2.1
ipython==8.30.0